+++ packages/react-native/src/pages/Signup/Nickname.tsx
