+++ packages/react-native/src/pages/Signup/NicknameProfile.tsx
